+++ config/routes.rb
@@ -112,7 +112,10 @@ Rails.application.routes.draw do
       get 'first', action: 'first', on: :collection
     end
 
-    get 'ical/externals' => 'ical#externals'
+    resources :i_calendar, only: %i[index create destroy] do
+      get 'events', on: :member
+      post 'sync', on: :member
+    end
 
     # for admin
     resources :trainings do